Update 'What is This' README.adoc
IPFS and Git do, in fact, have issues over WAN.  It turns out that IPFS does not play nicely with Git's storage structure and has trouble with Git's versioning.  Other options are being considered.  It should be noted that it is highly likely that this project's minimum requirements will be bumped up to an RPi4 w/ 4GB RAM.  This is unfortunate given the current processor shortage.